Feat: 메인페이지 나머지 버튼 컴포넌트 tailwind로 변경
