+++ src/app/components/year-selection.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { useShallow } from 'zustand/react/shallow';
+import { motion } from 'framer-motion';
+import classNames from 'classnames';
 
 import useYearStore from '@/app/stores/year';
 import useTeamStore from '@/app/stores/team';
 import { YEAR_LIST } from '@/app/const';
-
-import * as S from './styles';
 
 const YearSelection = () => {
   const [selectedYear, setYear, closePopup] = useYearStore(
@@ -20,7 +20,8 @@ const YearSelection = () => {
   };
 
   return (
-    <S.Container
+    <motion.div
+      className='absolute z-[1] grid w-[95vw] grid-cols-10 gap-2 rounded-sm border-2 bg-slate-200 p-5 tablet:w-[600px] laptop:w-[495px]'
       initial={{ scale: 0, opacity: 0 }}
       animate={{ scale: 1, opacity: 1 }}
       exit={{ scale: 0, opacity: 0 }}
@@ -31,18 +32,24 @@ const YearSelection = () => {
       }}
     >
       {YEAR_LIST.map((item, index) => (
-        <S.Item
+        <button
+          className={classNames(
+            'flex aspect-square items-center justify-center border-1 border-slate-900 bg-slate-100 text-[2.5vw] tablet:text-16 laptop:text-14',
+            {
+              'text-black hover:bg-[#e0a82433]': item && item !== selectedYear,
+              'bg-[#410] text-white': item && item === selectedYear,
+              'opacity-30': !item,
+            }
+          )}
           key={index}
           value={item ?? 0}
           onClick={onYearClick}
           disabled={!item}
-          $isDisabled={!item}
-          $isSelected={item === selectedYear}
         >
           {item ?? ''}
-        </S.Item>
+        </button>
       ))}
-    </S.Container>
+    </motion.div>
   );
 };
