+++ src/app/page.tsx
@@ -4,7 +4,7 @@ import { motion } from 'framer-motion';
 
 import MainTitle from './components/main-title';
 import Selection from './components/selection';
-import SubmitBtn from './components/submitBtn';
+import SubmitBtn from './components/submit-button';
 import Footer from './components/footer';
 
 export default function Home() {
